tambahdikit gan

+++ pengelola/data_semprop_diadmin.php
@@ -151,7 +151,7 @@ if($_SESSION['status'] == "login"){
           <td align='center'>$grade</td>
           <td align='center'>$key[status]</td>
           <td align='center'><a href='update_semrop_diadmin.php?nim=$key[nim]' role='button' class='btn btn-outline-primary'>EDIT</a>
-          <a href='delete_semprop_diadmin.php?nim=$key[nim]' role='button' class='btn btn-outline-primary'>DELETE</a></td>
+          <a href='delete_semprop_diadmin.php?nim=$key[nim]' role='button' class='btn btn-outline-primary' onclick='if(!confirm(`apakah anda yakin?`)){return false}'>DELETE</a></td>
           </tr>
          
         ";
